pyproject.toml
[build-system]
requires = ["setuptools>=72", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bb-run"
version = "1.0.0"
description = "Run Bitbucket Pipelines locally"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Karl Hill", email = "karlhillx@gmail.com" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "PyYAML>=6.0,<7",
]

[project.optional-dependencies]
# CI installs ``.[dev]``; pytest-cov supplies ``--cov`` / ``--cov-report``.
test = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/karlhillx/bb-run"
Repository = "https://github.com/karlhillx/bb-run"
Issues = "https://github.com/karlhillx/bb-run/issues"

[project.scripts]
bb-run = "bbrun.cli:main"

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
where = ["."]
include = ["bbrun*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py312"
line-length = 100
